fix: bookdrop -> bookDrop 패키지명 변경
