+++ src/main/java/com/t1/bookDrop/BookDropApplication.java
@@ -1,0 +1,13 @@
+package com.t1.bookDrop;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BookDropApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BookDropApplication.class, args);
+	}
+
+}
