+++ src/main/java/com/t1/bookDrop/controller/AuthController.java
@@ -1,0 +1,37 @@
+package com.t1.bookDrop.controller;
+
+import com.t1.bookDrop.aop.annotation.ValidAspect;
+import com.t1.bookDrop.dto.SigninReqDto;
+import com.t1.bookDrop.dto.SignupReqDto;
+import com.t1.bookDrop.service.AuthService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import javax.validation.Valid;
+
+@RestController
+@RequestMapping("/auth")
+public class AuthController {
+
+    @Autowired
+    private AuthService authService;
+
+    @ValidAspect
+    @PostMapping("/signup")
+    public ResponseEntity<?> signup(
+            @Valid @RequestBody SignupReqDto signupReqDto, BindingResult bindingResult
+    ) {
+        authService.signup(signupReqDto);
+        return ResponseEntity.created(null).body(true);
+    }
+
+    @PostMapping("/signin")
+    public ResponseEntity<?> signin (@RequestBody SigninReqDto signinReqDto) {
+        return ResponseEntity.ok(authService.signin(signinReqDto));
+    }
+}
