+++ src/main/java/com/t1/bookDrop/controller/advice/ExceptionControllerAdvice.java
@@ -1,0 +1,33 @@
+package com.t1.bookDrop.controller.advice;
+
+import com.t1.bookDrop.exception.SaveException;
+import com.t1.bookDrop.exception.ValidException;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ExceptionControllerAdvice {
+
+    @ExceptionHandler(SaveException.class)
+    public ResponseEntity<?> saveException(SaveException e) {
+        return ResponseEntity.internalServerError().body(e.getMessage());
+    }
+
+    @ExceptionHandler(ValidException.class)
+    public ResponseEntity<?> validException(ValidException e) {
+        return ResponseEntity.badRequest().body(e.getErrorMap());
+    }
+
+    @ExceptionHandler(UsernameNotFoundException.class)
+    public ResponseEntity<?> usernameNotFoundException(UsernameNotFoundException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+
+    @ExceptionHandler(BadCredentialsException.class)
+    public ResponseEntity<?> badCredentialsException(BadCredentialsException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+}
