+++ src/main/java/com/t1/bookDrop/dto/AdminSigninDto.java
@@ -1,0 +1,9 @@
+package com.t1.bookDrop.dto;
+
+import lombok.Data;
+
+@Data
+public class AdminSigninDto {
+    private String username;
+    private String password;
+}
