+++ src/main/java/com/t1/bookDrop/dto/SigninReqDto.java
@@ -1,0 +1,9 @@
+package com.t1.bookDrop.dto;
+
+import lombok.Data;
+
+@Data
+public class SigninReqDto {
+    private String username;
+    private String password;
+}
