+++ src/main/java/com/t1/bookDrop/entity/Admin.java
@@ -1,0 +1,16 @@
+package com.t1.bookDrop.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class Admin {
+    private int adminId;
+    private String username;
+    private String password;
+}
