+++ src/main/java/com/t1/bookDrop/entity/Book.java
@@ -1,0 +1,25 @@
+package com.t1.bookDrop.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class Book {
+    private int bookId;
+    private String bookName;
+    private String authorName;
+    private String publisherName;
+    private String isbn;
+    private int categoryId;
+    private String coverImgUrl;
+    private LocalDateTime createDate;
+    private LocalDateTime updateDate;
+
+}
