--- src/main/java/com/t1/bookdrop/BookdropApplication.java
@@ -1,13 +1,0 @@
-package com.t1.bookdrop;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BookdropApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BookdropApplication.class, args);
-	}
-
-}
