--- src/main/java/com/t1/bookdrop/aop/annotation/ParamsPrintAspect.java
@@ -1,11 +1,0 @@
-package com.t1.bookdrop.aop.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD})
-public @interface ParamsPrintAspect {
-}
