--- src/main/java/com/t1/bookdrop/dto/AdminSigninDto.java
@@ -1,9 +1,0 @@
-package com.t1.bookdrop.dto;
-
-import lombok.Data;
-
-@Data
-public class AdminSigninDto {
-    private String username;
-    private String password;
-}
