--- src/main/java/com/t1/bookdrop/dto/SigninReqDto.java
@@ -1,9 +1,0 @@
-package com.t1.bookdrop.dto;
-
-import lombok.Data;
-
-@Data
-public class SigninReqDto {
-    private String username;
-    private String password;
-}
