--- src/main/java/com/t1/bookdrop/entity/Admin.java
@@ -1,16 +1,0 @@
-package com.t1.bookdrop.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-public class Admin {
-    private int adminId;
-    private String username;
-    private String password;
-}
