--- src/main/java/com/t1/bookdrop/exception/SaveException.java
@@ -1,8 +1,0 @@
-package com.t1.bookdrop.exception;
-
-public class SaveException extends RuntimeException{
-
-    public SaveException() {
-        super("데이터 저장 오류");
-    }
-}
