--- src/main/java/com/t1/bookdrop/exception/ValidException.java
@@ -1,16 +1,0 @@
-package com.t1.bookdrop.exception;
-
-import lombok.Getter;
-
-import java.util.Map;
-
-public class ValidException extends RuntimeException{
-
-    @Getter
-    Map<String, String> errorMap;
-
-    public ValidException(Map<String, String> errorMap) {
-        super("유효성 검사 오류");
-        this.errorMap = errorMap;
-    }
-}
