--- src/main/java/com/t1/bookdrop/repository/AdminMapper.java
@@ -1,9 +1,0 @@
-package com.t1.bookdrop.repository;
-
-import com.t1.bookdrop.entity.Admin;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface AdminMapper {
-    public Admin adminCheckByUsername(String username);
-}
