+++ src/main/resources/mappers/user.xml
@@ -4,8 +4,8 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.t1.bookdrop.repository.UserMapper">
-    <resultMap id="UserResultMap" type="com.t1.bookdrop.entity.User">
+<mapper namespace="com.t1.bookDrop.repository.UserMapper">
+    <resultMap id="UserResultMap" type="com.t1.bookDrop.entity.User">
         <id property="userId" column="user_id" />
         <result property="username" column="username" />
         <result property="password" column="password" />
@@ -20,7 +20,7 @@
         <result property="updateDate" column="update_date" />
     </resultMap>
 
-    <insert id="saveUser" parameterType="com.t1.bookdrop.entity.User" keyProperty="userId" useGeneratedKeys="true">
+    <insert id="saveUser" parameterType="com.t1.bookDrop.entity.User" keyProperty="userId" useGeneratedKeys="true">
         insert into user_tb
         values(0, #{username}, #{password}, #{name}, #{email}, #{phone}, #{address}, #{gender}, #{age}, #{birth}, now(), now())
     </insert>
