+++ src/test/java/com/t1/bookDrop/BookdropApplicationTests.java
@@ -1,0 +1,13 @@
+package com.t1.bookDrop;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BookdropApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
