--- src/test/java/com/t1/bookdrop/BookdropApplicationTests.java
@@ -1,13 +1,0 @@
-package com.t1.bookdrop;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BookdropApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
